refactor: remove ContentActionBar component and integrate its functionality directly into ContentHeader for improved code clarity and maintainability
